Replace Assert with Return
Node's native `assert()` can be very heavy, especially in a large project
with a source map as it creates a full stack trace on failure.

In practice, we were seeing negative responses to model.find() taking
up to 2 seconds to resolve.

Most assertions in this library are either for config or use Koa's `assert`
method, which uses a considerably lighter code-path.

Replacing with if `(!stored) { return }` is an effectively free
operation.